some error handling

+++ src/apiMethods.js
@@ -16,28 +16,39 @@ const baseUrl = "https://reqres.in/api";
 const API = {
   getUsers:(page, cb) => {
 
+    if(typeof page !== 'number'){
+      console.error('invalid getUsers page parameter type', page);
+      return;
+    }
+
     const reqURL = `${baseUrl}/users/?page=${page}`
   
     axios.get(reqURL)
       .then((res)=>{
-        console.log('res', res);
-        if(cb) cb(res.data)
+        // console.log('res', res);
+        if(cb) cb(res)
       })
       .catch((err)=>{
-        console.log('err', err);
+        console.error('err', err);
       })
   
   },
   deleteUser: (userID, cb) => {
+
+    if(typeof userID !== 'number'){
+      console.error('invalid deleteUser userID parameter type', userID);
+      return;
+    }
+    
     const reqURL = `${baseUrl}/users/${userID}`
   
     axios.delete(reqURL)
       .then((res)=>{
-        console.log('res', res);
+        // console.log('res', res);
         if(cb) cb(res)
       })
       .catch((err)=>{
-        console.log('err', err);
+        console.error('err', err);
       })
 
   }
